+++ CHANGELOG.md
@@ -2,8 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.34.2] – 2022-05-30
+## [1.34.3] – 2022-06-08
+
+### New
+
+- send `accessKey` header in api requests (specified in `config.network.accessKey`)
+
 ### Fixed
+
+- send headers in `info` api requests
+
+## [1.34.2] – 2022-05-30
+
+### Fixed
+
 - build process
 
 ## [1.34.1] – 2022-05-26